requirements.txt
Django==1.11
django-cors-headers==2.1.0
djangorestframework==3.7.7
Pillow==5.0.0
pytz==2017.3
